+++ app-sample/src/main/java/io/noties/markwon/app/samples/table/TableLinkifySample.java
@@ -3,8 +3,10 @@ package io.noties.markwon.app.samples.table;
 import io.noties.markwon.Markwon;
 import io.noties.markwon.app.sample.Tags;
 import io.noties.markwon.app.sample.ui.MarkwonTextViewSample;
+import io.noties.markwon.ext.tables.TableAwareMovementMethod;
 import io.noties.markwon.ext.tables.TablePlugin;
 import io.noties.markwon.linkify.LinkifyPlugin;
+import io.noties.markwon.movement.MovementMethodPlugin;
 import io.noties.markwon.sample.annotations.MarkwonArtifact;
 import io.noties.markwon.sample.annotations.MarkwonSampleInfo;
 
@@ -12,7 +14,7 @@ import io.noties.markwon.sample.annotations.MarkwonSampleInfo;
   id = "20200702135739",
   title = "Linkify table",
   description = "Automatically linkify markdown content " +
-    "including content inside tables",
+    "including content inside tables, handle clicks inside tables",
   artifacts = {MarkwonArtifact.EXT_TABLES, MarkwonArtifact.LINKIFY},
   tags = {Tags.links}
 )
@@ -23,10 +25,11 @@ public class TableLinkifySample extends MarkwonTextViewSample {
       "| HEADER | HEADER | HEADER |\n" +
       "|:----:|:----:|:----:|\n" +
       "|   测试  |   测试   |   测试   |\n" +
-      "|  测试  |   测试   |  测测测12345试测试测试   |\n" +
+      "|   测试  |   测试   |  测测测12345试测试测试   |\n" +
       "|   测试  |   测试   |   123445   |\n" +
       "|   测试  |   测试   |   (650) 555-1212   |\n" +
-      "|   测试  |   测试   |   [link](#)   |\n" +
+      "|   测试  |   测试   |   mail@ma.il   |\n" +
+      "|   测试  |   测试   |   some text that goes here is very very very very important [link](https://noties.io/Markwon)   |\n" +
       "\n" +
       "测试\n" +
       "\n" +
@@ -35,6 +38,8 @@ public class TableLinkifySample extends MarkwonTextViewSample {
     final Markwon markwon = Markwon.builder(context)
       .usePlugin(LinkifyPlugin.create())
       .usePlugin(TablePlugin.create(context))
+      // use TableAwareLinkMovementMethod to handle clicks inside tables
+      .usePlugin(MovementMethodPlugin.create(TableAwareMovementMethod.create()))
       .build();
 
     markwon.setMarkdown(textView, md);